requirements.txt
selenium
Django==2.0.3
django-jquery==3.1.0
djangorestframework==3.8.2
Jinja2==2.10
MarkupSafe==1.0
ObjDict==0.4.4
pytz==2018.3
virtualenv==15.2.0
